+++ save_test.go
@@ -73,7 +73,7 @@ func TestSave_WithModifiedAt(t *testing.T) {
 	objSaved := test.TestStructWithData()
 	err := testCRUD.Save(context.Background(), objSaved, SaveOptions{
 		ModifiedBy: 4,
-		ModifiedAt: uint64(time.Now().Unix()),
+		ModifiedAt: time.Now().Unix(),
 	})
 	if err != nil {
 		t.Fatalf("Save failed to insert struct to the table: %s", err.(*CRUDError).Op)
@@ -94,7 +94,7 @@ func TestSave_WithModifiedAt(t *testing.T) {
 	// Now, update the object in the database
 	objSaved.PrimaryEmail = "primary1@example.com"
 
-	modifiedAt := uint64(time.Now().Unix())
+	modifiedAt := time.Now().Unix()
 	modifiedBy := uint64(5)
 
 	err = testCRUD.Save(context.Background(), objSaved, SaveOptions{
